requirements.txt
django
python-dotenv
openai
matplotlib
plotly
scipy
scikit-learn
numpy
pandas
whitenoise
azure-cognitiveservices-speech
PyPDF2
